Incorrect change to IOContext. Committing for easy diff.

+++ include/partition_and_pq.h
@@ -16,7 +16,7 @@ void gen_random_slice(const std::string base_file,
                       const std::string output_prefix, float sampling_rate);
 
 template<typename T>
-void gen_random_slice(const std::string data_file, float p_val,
+void gen_random_slice(const std::string data_file, double p_val,
                       float *&sampled_data, size_t &slice_size, size_t &ndims);
 
 template<typename T>
